pcap-store/requirements.txt
confluent-kafka==2.3.0
orjson==3.9.15
psycopg2-binary==2.9.9
minio==7.2.7
